fix(server): share proxy IP helper and check IP limits first
Extract trusted_client_ip into net.rs for auth and rate limiting, and
enforce IP buckets before session buckets to avoid consuming user quota
on IP-throttled rejections.
